Use Java 8 standards, up-to-date Spring standards, and AssertJ instead of old-style JUnit. And other fixes :)

+++ src/main/java/com/exo/ecommerce/domain/item/Item.java
@@ -6,11 +6,12 @@ public class Item {
     private String name;
     private String description;
     private Integer remainingInStock;
-    private Float price;
+    private Double price;
 
-    public Item() {}
+    public Item() {
+    }
 
-    public Item(Long id, String name, String description, Integer remainingInStock, Float price) {
+    public Item(Long id, String name, String description, Integer remainingInStock, Double price) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -34,7 +35,7 @@ public class Item {
         return remainingInStock;
     }
 
-    public Float getPrice() {
+    public Double getPrice() {
         return price;
     }
